Revert "docs: ERCOT NORTH_C real-data benchmark report (Sprint 15)"
