+++ thoughts/shared/templates/benchmark-iteration-report.md
@@ -8,61 +8,10 @@
 | Benchmark | `{benchmark_name}` (e.g., `predictive_energy`) |
 | Date | `{YYYY-MM-DD}` |
 | Commit | `{short_hash}` on branch `{branch_name}` |
-| Dataset ID | `{dataset_id}` (e.g., `ercot_north_c_dfw_2022_2024`) |
 | Dataset | `{path_or_description}` |
 | Dataset Rows | `{N}` |
 | Split | `{train_frac}/{val_frac}/{test_frac}` (e.g., `60/20/20`) |
 | Reporter | `{agent_id or human_name}` |
-| Timezone | `{timezone_of_saved_dataset}` (recommended: `UTC`) |
-| Target Series | `{area_or_series_name}` |
-| Weather Source | `{station_name_or_description}` |
-
-## Dataset Provenance
-
-### Source Inputs
-
-| Input | Value |
-|-------|-------|
-| Load Source | `{official_source_name}` |
-| Load URL | `{official_url}` |
-| Load Series | `{series_name}` |
-| Weather Source | `{official_source_name}` |
-| Weather URL | `{official_url}` |
-| Weather Station | `{station_id} - {station_name}` |
-| Raw Coverage | `{start_timestamp}` to `{end_timestamp}` |
-| Prepared Dataset Path | `{local_path_to_prepared_csv_or_parquet}` |
-| Prepared File Size | `{size}` |
-| Prepared File SHA256 | `{sha256_or_n/a}` |
-
-### Preparation Notes
-
-1. `{How timestamps were normalized and converted to the saved timezone}`
-2. `{How load and weather were joined}`
-3. `{How duplicate timestamps were handled}`
-4. `{How missing temperature/humidity values were handled}`
-
-## Dataset QA
-
-### Structural Checks
-
-| Check | Result | Detail |
-|-------|--------|--------|
-| Required columns present | `{PASS/FAIL}` | |
-| Timestamps parse cleanly | `{PASS/FAIL}` | |
-| Timestamps are unique | `{PASS/FAIL}` | |
-| Timestamps monotonic after sort | `{PASS/FAIL}` | |
-| Single-series dataset | `{PASS/FAIL}` | |
-| No missing `target_load` | `{PASS/FAIL}` | |
-| No missing `temperature` | `{PASS/FAIL}` | |
-| Saved timezone is explicit | `{PASS/FAIL}` | |
-
-### Split Boundaries
-
-| Partition | Rows | Start | End |
-|-----------|------|-------|-----|
-| Train | `{N}` | `{timestamp}` | `{timestamp}` |
-| Validation | `{N}` | `{timestamp}` | `{timestamp}` |
-| Test | `{N}` | `{timestamp}` | `{timestamp}` |
 
 ## Run Configuration
 
@@ -130,22 +79,13 @@ W = wins, L = losses, T = ties (within 1% relative)
 
 ### Selected Model Distribution
 
-| Strategy | ridge | rf | gbm |
+| Strategy | ridge | random_forest | hist_gbm |
 |----------|-------|---------------|----------|
 | random | `{count}` | `{count}` | `{count}` |
 | surrogate_only | `{count}` | `{count}` | `{count}` |
 | causal | `{count}` | `{count}` | `{count}` |
 
 Counts are across all seeds and budgets for the best-selected configuration per run.
-
-### Artifact Inventory
-
-| Artifact | Path | Notes |
-|----------|------|-------|
-| Smoke JSON | `{path}` | `{optional}` |
-| Full Results JSON | `{path}` | |
-| Summary CSV | `{path}` | |
-| Report Markdown | `{path}` | |
 
 ## Interpretation
 
@@ -197,6 +137,3 @@ uv run python scripts/energy_predictive_benchmark.py \
 ```
 
 Commit: `{full_hash}`
-
-If dataset preparation was part of this iteration, include the exact prep command
-or script path here as well.
